+++ dapt/db/delimited_file.py
@@ -5,7 +5,73 @@
 Delimited file
 ==============
 
-Create a CSV database which can be used by param to get and run param sets.  
+This class uses a local `delimited file
+<https://en.wikipedia.org/wiki/Delimiter-separated_values>`_ (e.g. CSV, TSV) as a database for
+parameter testing.  Delimited files represent a table where the row is a line in the file, and
+each column is seporated by a delimiter.  The delimiter is often a comma (comma-seporated file)
+or a tab (tab-seporated file).  A CSV file might look like this:
+
+::
+
+    id,start-time,end-time,status,a,b,c
+    t1,2019-09-06 17:23,2019-09-06 17:36,finished,2,4,6
+    t2,,,,10,10,
+    t3,,,,10,-10,
+
+and represent a table that looks like this:
+
++----+------------------+------------------+----------+----+-----+---+
+| id | start-time       | end-time         | status   | a  | b   | c |
++----+------------------+------------------+----------+----+-----+---+
+| t1 | 2019-09-06 17:23 | 2019-09-06 17:36 | finished | 2  | 4   | 6 |
++----+------------------+------------------+----------+----+-----+---+
+| t2 |                  |                  |          | 10 | 10  |   |
++----+------------------+------------------+----------+----+-----+---+
+| t3 |                  |                  |          | 10 | -10 |   |
++----+------------------+------------------+----------+----+-----+---+
+
+Because these files are stored on the users computer, there is no way for a team to work on
+the parameter set distributively (without manually dividing the parameter sets up).
+
+Delimited files can have a header which gives the columns names.  The header is the first row
+of the table.  Headers must be included with DAPT's ``Delimited_file`` class.
+
+DAPT provides a method named ``sample_db()`` which creates the sample CSV above.  You can create
+this file by running that method and then use the ``Delimited_file`` class with it.
+
+    >>> db = dapt.tools.sample_db(file_name='sample_db.csv', delimiter=',')
+    >>> db.fields()
+    ['id', 'start-time', 'end-time', 'status', 'a', 'b', 'c']
+
+
+Config
+------
+
+Delimited file can accept a :ref:`config` class.  The values listed in the table below are
+the same attributes used to instantiate the class.  These values should be placed inside
+a JSON object named ``delimited-file.`` 
+
++---------------------------+----------------------------------------------------------------+
+| Fields                    | Description                                                    |
++===========================+================================================================+
+| ``path`` (str)            | The path, from the execution directory, to the delimited file. |
++---------------------------+----------------------------------------------------------------+
+| ``delimiter`` (str)       | How the columns of the file are seporated.                     |
++---------------------------+----------------------------------------------------------------+
+
+The default configuration looks like this:
+
+.. code-block:: JSON
+    :caption: Sample JSON configuration for ``Delimited_file``
+    :name: example-delimited-file-config
+
+    {
+        "delimited-file" : {
+            "path" : "parameters.csv", 
+            "delimiter" : ","
+        }
+    }
+
 """
 
 import csv
@@ -20,34 +86,88 @@ class Delimited_file(base.Database):
     """
     An interface for accessing and setting paramater set data.  
 
-    Args:
-        path (string): path to delimited file file
+    Keyword args:
+        path (str): path to delimited file file
+        delimiter (str): the delimiter of the CSV.  ``,`` by default.
+        config (``Config`` object): an Config instance
     """
         
-    def __init__(self, path, delimiter=','):
+    def __init__(self, *args, **kwargs):
         
         super().__init__()
         
-        self.path = path
-        self.delimiter = delimiter
+        self.config = None
+        self.path = None
+        self.delimiter = ','
+
+        if len(kwargs) == 0 and len(args) == 0:
+            raise ValueError("You must provide a Config object or path to the CSV.")
+        if 'config' in kwargs:
+            self.config = kwargs['config']
+            if self.config.has_value('delimited-file'):
+                if self.config.has_value(['delimited-file', 'path']):
+                    self.path = self.config['delimited-file']['path']
+                if self.config.has_value(['delimited-file', 'delimiter']):
+                    self.delimiter = self.config['delimited-file']['delimiter']
+        
+        # Check for values given as args
+        if len(args) > 0:
+            self.path = args[0]
+            if len(args) > 1:
+                self.delimiter = args[1]
+
+        # Check for values given as kwargs
+        if 'path' in kwargs:
+            self.path = kwargs['path']
+        if 'delimiter' in kwargs:
+            self.delimiter = kwargs['delimiter']
+
 
     def connect(self):
         """
-        This method isn't required for Delimeted files as there is nothing to connect to.
-        It will return True regardless.
+        The method used to connect to the database and log the user in.  Some databases won't
+        need to use the connect method, but it should be called regardless to prevent problems.
 
         Returns:
             True if the database connected successfully and False otherwise.
         """
 
-        return True
+        return self.connected()
+
+    def connected(self):
+        """
+        Check to see if the API is connected to the server and working.
+
+        Returns:
+            True if the API is connected to the server and False otherwise.
+        """
+
+        if os.path.exists(self.path) and os.path.isfile(self.path):
+            try:
+                with open(self.path, 'r') as csvfile:
+                    csv.DictReader(csvfile, delimiter=self.delimiter)
+                
+            except Exception as e:
+                _log.warning("Cannot read the delimited file: %s" % str(e))
+                return False
+            
+            if not os.access(self.path, os.W_OK):
+                _log.warning("Cannot write to delimited file.")
+                return False
+                
+            return True
+        else:
+            _log.warning("Delimited file does not exist.")
+
+        return False
 
     def get_table(self):
         """
         Get the table from the database.
 
         Returns:
-            An array with each element being a dictionary of the key-value pairs for the row in the database.
+            An array with each element being a dictionary of the key-value pairs for the row
+            in the database.
         """
 
         sheet = []
@@ -63,11 +183,13 @@ class Delimited_file(base.Database):
         
         Returns:
             Array of strings with each element being a field (order is preserved if possible)
+            or ``None`` if the file is empty.
         """
 
         with open(self.path, 'r') as csvfile:
             reader = csv.reader(csvfile, delimiter=self.delimiter)
-            return next(reader)
+
+            return next(reader, None)
 
     def update_row(self, row_index, values):
         """